mv action to one file
